# openapi/dazat-api.yaml
openapi: 3.1.0
info:
  title: DAZAT Mobility API
  version: 0.0.6
  description: Engineering Phase 0.6 API. Server/domain owners remain authoritative for identity, Booking, Dispatch, Journey telemetry, governed changes, Safety and completion evidence.
servers:
  - url: http://localhost:3001
paths:
...
  /v1/bookings/{bookingId}/journey/acknowledge:
    post:
      summary: Assigned Driver acknowledges work and creates the authoritative EN_ROUTE Journey/first leg.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/BookingId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': { description: Active assignment materialised as Journey and advanced to EN_ROUTE. }
        '409': { description: Assignment or Booking state changed. }
  /v1/bookings/{bookingId}/journey:
    get:
      summary: Read the authorised Rider/Driver Journey projection for reconciliation.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200': { description: Current Journey projection with telemetry freshness and hold state. }
        '404': { description: Assigned Driver has not yet acknowledged and created the Journey. }
  /v1/journeys/{journeyId}/location-observations:
    post:
      summary: Ingest an append-only, purpose-scoped Driver location observation with source, time, accuracy and confidence.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/JourneyId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DriverLocationObservationRequest' }
      responses:
        '202': { description: Observation durably recorded; stale/degraded evidence remains explicitly labelled. }
  /v1/journeys/{journeyId}/arrived:
    post:
      summary: Mark arrival only after fresh, accurate, confident evidence falls inside the configured pickup radius.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/JourneyId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': { description: Immutable arrival evidence accepted and Journey advanced to ARRIVED. }
        '409': { description: Location evidence is missing, stale, degraded or outside the radius. }
  /v1/journeys/{journeyId}/ridecheck/start:
    post:
      summary: Passenger starts protected RideCheck; the PIN is returned once and never persisted in plaintext.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/JourneyId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201': { description: Salted-verifier RideCheck session created and challenge returned once. }
        '409': { description: Journey is not in ARRIVED or a protected hold exists. }
  /v1/journeys/{journeyId}/ridecheck/verify:
    post:
      summary: Assigned Driver submits passenger PIN; mismatch is recorded without becoming a misconduct finding.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/JourneyId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rideCheckSessionId, code]
              properties:
                rideCheckSessionId: { type: string, format: uuid }
                code: { type: string, pattern: '^\\d{6}$' }
      responses:
        '200': { description: Passenger/Driver/vehicle pairing verified. }
        '409': { description: Mismatch, expiry or exhausted attempts; protected start remains blocked. }
  /v1/journeys/{journeyId}/start:
    post:
      summary: Start only after verified RideCheck, active eligible assignment, fresh pickup evidence and no active hold.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/JourneyId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': { description: Backend-validated JourneyStarted transition committed. }
        '409': { description: One or more protected-start preconditions failed. }
...
  /v1/journeys/{journeyId}/active:
    get:
      summary: Alias for the authoritative active Journey reconciliation snapshot.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/JourneyId'
      responses:
        '200': { description: Authoritative snapshot; reconnecting clients replace speculative local state. }
  /v1/journeys/{journeyId}/telemetry/location:
    post:
      summary: Append active Journey telemetry with confidence and movement-plausibility classification.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/JourneyId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DriverLocationObservationRequest' }
      responses:
        '202': { description: Observation persisted; impossible jumps and out-of-order samples remain explicit degraded evidence. }
        '409': { description: Journey is not active or observation identity conflicts. }
  /v1/journeys/{journeyId}/stop-requests:
    post:
      summary: Request an additional stop without mutating the active route prematurely.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/JourneyId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StopRequest' }
      responses:
        '202': { description: Request is PENDING_POLICY_REVIEW; pricing and Driver acknowledgement are not fabricated. }
        '409': { description: Expected Journey version is stale or Journey is inactive. }
  /v1/journeys/{journeyId}/route-change-requests:
    post:
      summary: Record a governed stop/destination request with an immutable requested location snapshot.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/JourneyId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RouteChangeRequest' }
      responses:
        '202': { description: Request recorded but routeChanged remains false pending policy workflow. }
  /v1/journeys/{journeyId}/arriving:
    post:
      summary: Enter ARRIVING only with fresh, plausible destination-approach evidence.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/JourneyId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': { description: Destination approach and completion-requirement snapshot committed. }
        '409': { description: Destination evidence is missing, stale, degraded, implausible or outside policy radius. }
  /v1/journeys/{journeyId}/complete:
    post:
      summary: Complete only after all authoritative destination, assignment, hold, continuity and handover guards pass.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/JourneyId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': { description: Journey/Booking/assignment completed and Driver returned AVAILABLE; paymentInitiated is false. }
        '409': { description: One or more governed completion requirements failed. }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 200 }
    BookingId:
      name: bookingId
      in: path
      required: true
      schema: { type: string, format: uuid }
    JourneyId:
      name: journeyId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    LocationInput:
      type: object
      required: [latitude, longitude, displayLabel]
      properties:
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        displayLabel: { type: string }
...
    DriverLocationObservationRequest:
      type: object
      required: [clientObservationId, latitude, longitude, observedAt, source, accuracyMetres, confidence]
      properties:
        clientObservationId: { type: string, format: uuid }
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        observedAt: { type: string, format: date-time }
        source: { type: string, enum: [DEVICE_GPS, FLEET_TELEMATICS] }
        accuracyMetres: { type: number, minimum: 0 }
        confidence: { type: number, minimum: 0, maximum: 1 }
    StopRequest:
      type: object
      required: [location, expectedJourneyVersion, reasonCode]
      properties:
        location: { $ref: '#/components/schemas/LocationInput' }
        expectedJourneyVersion: { type: integer, minimum: 1 }
        reasonCode: { type: string, enum: [PASSENGER_REQUEST, ACCESSIBILITY_NEED, OPERATIONAL_NEED, OTHER] }
    RouteChangeRequest:
      allOf:
        - $ref: '#/components/schemas/StopRequest'
        - type: object
          required: [requestType]
          properties:
            requestType: { type: string, enum: [ADD_STOP, CHANGE_DESTINATION] }
...